# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Campus Resource Management and Booking System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: Search resources with availability
      parameters:
        - name: category_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: min_capacity
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: string
        - name: start_time
          in: query
          schema:
            type: string
        - name: end_time
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Resources with availability
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourceWithAvailability"
...
  /resources/{id}:
    get:
      operationId: getResource
      tags: [resources]
      summary: Resource detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resource detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceDetail"
  /resources/{id}/availability:
    get:
      operationId: getResourceAvailability
      tags: [resources]
      summary: Get availability for a specific date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlots"

  # BOOKINGS
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    ResourceDetail:
      type: object
      properties:
        resource_id:
          type: integer
        resource_name:
          type: string
        capacity:
          type: integer
        location:
          type: ["string", "null"]
        status:
          type: string
        features:
          type: object
        category_id:
          type: integer
        category_name:
          type: string
        approval_steps:
          type: integer
        manager_id:
          type: ["integer", "null"]
        manager_name:
          type: ["string", "null"]
        department_id:
          type: ["integer", "null"]
        department_name:
          type: ["string", "null"]
        unavailability:
          type: array
          items:
            $ref: "#/components/schemas/ResourceUnavailability"
      required: [resource_id, resource_name, capacity, status, features, category_id, category_name, approval_steps]
    ResourceWithAvailability:
      type: object
      properties:
        resource_id:
          type: integer
        resource_name:
          type: string
        capacity:
          type: integer
        location:
          type: ["string", "null"]
        status:
          type: string
        features:
          type: object
        category_id:
          type: integer
        category_name:
          type: string
        approval_steps:
          type: integer
        manager_name:
          type: ["string", "null"]
        is_available:
          type: boolean
      required: [resource_id, resource_name, capacity, status, features, category_id, category_name, approval_steps, is_available]
...
    ResourceUnavailability:
      type: object
      properties:
        unavail_id:
          type: integer
        resource_id:
          type: integer
        day_of_week:
          type: ["integer", "null"]
        start_time:
          type: string
        end_time:
          type: string
        label:
          type: ["string", "null"]
      required: [unavail_id, resource_id, start_time, end_time]
...
    AvailabilitySlots:
      type: object
      properties:
        busy_slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
        free_slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
      required: [busy_slots, free_slots]
    TimeSlot:
      type: object
      properties:
        start_time:
          type: string
        end_time:
          type: string
        booking_id:
          type: ["integer", "null"]
        purpose:
          type: ["string", "null"]
      required: [start_time, end_time]
...