# lib/api-zod/src/generated/types/resourceDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Resource Management and Booking System API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceDetailFeatures } from "./resourceDetailFeatures";
import type { ResourceUnavailability } from "./resourceUnavailability";

export interface ResourceDetail {
  resource_id: number;
  resource_name: string;
  capacity: number;
  /** @nullable */
  location?: string | null;
  status: string;
  features: ResourceDetailFeatures;
  category_id: number;
  category_name: string;
  approval_steps: number;
  /** @nullable */
  manager_id?: number | null;
  /** @nullable */
  manager_name?: string | null;
  /** @nullable */
  department_id?: number | null;
  /** @nullable */
  department_name?: string | null;
  unavailability?: ResourceUnavailability[];
}