# lib/api-zod/src/generated/types/resourceUnavailability.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Resource Management and Booking System API
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceUnavailability {
  unavail_id: number;
  resource_id: number;
  /** @nullable */
  day_of_week?: number | null;
  start_time: string;
  end_time: string;
  /** @nullable */
  label?: string | null;
}