# lib/api-zod/src/generated/types/resourceWithAvailability.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Resource Management and Booking System API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceWithAvailabilityFeatures } from "./resourceWithAvailabilityFeatures";

export interface ResourceWithAvailability {
  resource_id: number;
  resource_name: string;
  capacity: number;
  /** @nullable */
  location?: string | null;
  status: string;
  features: ResourceWithAvailabilityFeatures;
  category_id: number;
  category_name: string;
  approval_steps: number;
  /** @nullable */
  manager_name?: string | null;
  is_available: boolean;
}